archival: fallback flush of last_clean_at
Locations that upload the manifest should also
consider whether they should flush the clean_at
offset in the state machine: this may generate
an extra raft write, but it it is necessary to
do this write eventually to avoid holding up
log prefix truncation if the partition is idle
and doesn't have any incoming writes to prompt
manifest updates in the upload loop.

This time-based condition covers two cases:
- Bugs if we update the manifest in the background
  but fail to flush the clean_at offset.
- Any failure condition writing the clean_at offset
  that doesn't cause leadership to be lost: errors
  like this could result in setting a projected
  clean_at but never flushing it, if we didn't have
  this fallback case.

The naming of methods is changed: previously
maybe_flush_manifest_clean_offset was always flushing
if there was anything to flush: the maybe_ is dropped
from this method.  A new maybe_flush_manifest_clean_offset
method is introduced which uses a time condition to only issue
a flush if one has not been done for some time.

The complexity of this change reflects the general complexity
of the upload loop logic that results from sometimes inlining
mark_clean operations with add_segments, vs. sometimes
doing them independently, and the complexity from
sometimes uploading manifests in parallel with data segments,
and sometimes separately: these are both performance-oriented
behaviors that result in lots more branches in the code.  This
can and should be refactored in future.