Restrict replicate batcher to 1 flush in flight
AppendEntry requests to partition replicas flow through the replicate
batcher, which may accumulate multiple entries in an "item cache" before
sending them all during a flush() operation.

Each time cache entries are enqueued, a background flush is triggered
which will eventually fulfill the promise associated with each item.
The flush task triggered after queuing an item is not necessarily the
task that flushes it: an flush enqueued by an earlier flush may also
flush the item.

If there are a high rate of append entries requests, the flush tasks
can accumulate at a faster rate than they are drained. The actual
flush operations are occurring with reasonable frequency and since each
flush drains the entire item cache, there is no backlog of unsent
items, only a backlog of tasks and the associated error-handling
continuation. After the load stops these tasks quickly drain and the
flush is a no-op in general since the item cache is empty.

However, if the load does not stop in time, an out of memory can
occur as the tasks grow indefinitely.

This change fixes this by ensuring that only one flush task is
outstanding at any time, since the additional flushes are redundant.

Fixes #9533.