cloud_storage: Support DeleteRecords in 'archive'
To support DeleteRecords in 'archive' section of the log the truncate
command should be able to accept offsets lower than start_kafka_offset
of the manifest. Also, the manifest has to store not only
archive_start_offset but also a corresponding archive_start_delta value
to be able to compute start kafka offset of the 'archive' section of
the log. This value is needed to sanity check the offset applied by
DeleteRecords request.