vlog: add vlog_error macro for structured errors
While we do not do structured logging in general, for
specific error cases it is useful to assign unique
error codes.  In future when we do structured logging
these codes will carry forward.

Error codes are defined centrally for ease of reference,
and to make it easy to avoid colliding codes between
different logging locations.