net/transport: track shutdown of current socket
This commit tweaks the `base_transport` class to track whether
`shutdown` was called on the current socket. This is required because
the lifetime of `base_transport` does not match the lifetime of a single
connection. Previously, this was done implicitly through the
`base_transport::_fd` pointer.

The reason for tracking this property is to keep the pointer to the
underlying connection, `_fd`, alive for longer. A follow-up commit will
make use of this.