net/transport: expose wait_input_shutdown
Recent versions of seastar expose a `wait_for_input_shutdown` on
connections. This method can be used to wait until the peer disconnects.

This is particularly useful when dealing with TLS connections. Seastar
spawns a future in the background which waits for the peer to gracefully
terminate the connection. If the peer does not do this in a timely
manner, this can lead to memory leaks during shutdown.

This commit simply introduces the method. A future commit wires it up.