cloud_storage_clients/abs: wait for conn shutdown
This commit updates the ABS client to _wait_ for the shutdown of its
underlying connection. A lot of times, Azure will not terminate TLS
connections promptly when prompted to do so. This, coupled with the fact
that Seastar spawns a background fiber to wait on the final ACK, leads
to memory leaks upon shutdown.