client_pool: stop all clients when closing service
Previously, the client pool did not stop all of its clients when the
inner sharded service was stopped. Generally, this was fine since the
connections where being shutdown explicitly via `shutdown_connections`
on tear-down. However, `shutdown_connections` does not close the gate,
so in theory new connections could be created in the meantime.

Explicitly calling `stop` is the safe thing do here. For ABS clients,
the `stop` call also waits for the peer to close the connection in order
to work around Azure being slow to close TLS connections.