admin: interface for spawning busy loops
This adds an admin endpoint to kick off some number of fibers to run
busy loops (i.e. loops that just hog CPU time without doing anything
meaningful). The goal is to put redpanda in a state that emulates heavy
load.

The loops aren't too fancy: they just loop for some number of times as
input by the admin endpoint, and periodically yield. Currently only one
loop is allowed to run at a time per shard and it loops a fixed number
of times before yielding. In the future we may want to extend this, e.g.
by running potentially multiple loops with different scheduling groups,
or with randomized spins.