storage: apply conservative size-based retention
When deleting segments for size-based retention policy a segment should
not be deleted if it would delete more than the target retention size.
This is the same semantics as kafka, and avoids a problematic case where
we might delete a large segment, say 1 GB, just to reclaim a few
additional bytes according to the policy. This could lead to unexpected
behavior.

Fixes: https://github.com/redpanda-data/redpanda/issues/9526

Signed-off-by: Noah Watkins <noahwatkins@gmail.com>