raft: avoid caching batches when only one node is recovering
In cases where one follower falls far behind(like when a node goes down)
it is often the case that we are reading large amounts of old batches only
once. If these reads are placed in the batch cache then newer, more
likely to be consumed, batches will be evicted. This can cause a
noticable increase in batch cache misses during recovery.

This commit avoids this problem by not adding read batches to the cache
in cases where only one follower is recovering. In the cases where more
than one follower is recovering its no longer a read-once workload and
it may be optimal to put read batches in the cache.