rptest: fix incorrect flag usage
rpk.sasl_create_user always used --sasl-mechanism, which is actually the
flag for the Kafka protocol. Not installing the Kafka flags under all of
`rpk acl` breaks any test that used sasl_create_user.

The default mechanism for the proper flag, --mechanism, is
SCRAM-SHA-256. That matches the mechanism defined for
SUPERUSER_CREDENTIALS, which is why the test always previously worked.

We now use the proper flag