rpk: strip rpk flags before execing plugins
This is even more important in the context of -X, which we expect to be
somewhat hardcoded.

Previously, if you ran a plugin with `-v`, the -v would be passed to the
plugin and the plugin would fail if it did not have that command.

Now, we strip all rpk specific flags before execing plugins.

This is done in all places where DisableFlagParsing is set to true.