ducktape: transactions multiple tm partitions upgrade test
Add upgrade test to new version of redpanda that supports multiple
tm partitions.
New version contains new fence_batch and snapshot versions.
So it is impossible to downgrade from new version to previous
We test that flag of uasge new versions of fetch batch and sanpshot
works correctly.
Check that it is impossible to downgrade cluster from new version