archival: sync manifest before GC'ing segments
The order of operations needs to be:
1. Move segments to 'replaced' list
2. Upload manifest so that anyone reading the bucket no longer
   sees the to-be-removed segment as readable.
3. Actually delete the segments

Previously we were doing 1,3,2 instead of 1,2,3.

This is most obvious when doing an end-of-test validation
of the bucket, because the validation tool sees what a read
replica cluster would see: a manifest that points to segments
that no longer exist.

There are some cases where we could avoid doing the manifest
upload even if it is dirty, if it had been uploaded since the
segment moved to the replaced list, but this would require
tracking a segment-by-segment record of at what stm offset
it was moved into the list, and would only rarely yield
any perf benefit, as manifests are uploaded quite promptly
under normal operation: the chances are that when we do
an infrequent housekeeping pass, the manifest is already
uploaded and this new upload call is a no-op.

Fixes https://github.com/redpanda-data/redpanda/issues/10086