cloud_storage: Add segment chunk objects
A segment chunk and a chunks API are added. The segment chunk represents
a portion of the segment file, which may or may not be hydrated, IE
downloaded to disk.

Each chunk manages its own file handle. The handle is opened when the
chunk is hydrated, and is kept open until the chunk is "trimmed" by the
segment_chunks API.

A chunk may or may not be required by readers. We periodically trim the
hydrated chunks which are not currently in use by any readers, by taking
into account some metrics.

Each chunk has some associated metrics like:

* The number of readers which will need this chunk after they are done
  reading their current chunk. A high count here represents a chunk
  which should not be evicted, because many readers will soon need it.
* The number of chunks after which the current chunk will be needed by
  readers, represented as a sum. For example if a chunk is required by
  reader A after 3 chunks, and the same chunk is required by reader B
  after 2 chunks, then it's score is 5. A lower score means a chunk is
  going to be required soon and should not be trimmed. A chunk with high
  score should be safer to trim compared to one with a lower score.

Chunks request hydration via the remote segment. The hydration process is
very similar to the hydration loop in the remote segment. The first
reader requesting hydration starts the process, and any subsequent readers
requesting the same chunk are added to a wait list.

On successful hydration but failed materialization we assume that the
chunk was evicted from cache, and try again after a short sleep, same as
the retry behavior in the remote segment.

The chunk file handle is shared with readers as a shared pointer, which
enables the trim process to figure out which chunks are unused as their
handles are "owned".