archival: remove redundant manifest upload after GC
The important upload is the one that happens inside
garbage_collect(), because we need to update the
manifest to avoid external readers (scrubbers and
read replicas) getting upset that they got a 404
reading a segment that's meant to exist.

The upload _after_ garbage collection only serves
to trim the 'segments' and/or 'replaced' vectors
in the remote copy of the manifest, which has no
logical impact.  We can rely on the next periodic
manifest upload (manifest_upload_interval) to
pick that up.

Followup to https://github.com/redpanda-data/redpanda/pull/10099