storage: introduce failure injection config types
This commit introduces a number of utility types that will be used to
implement failure injection within the storage layer. The most
significant part of this commit is the JSON schema which defines the
expected shape of the failure injection config.

Logic for converting from JSON to the new types is also included. Note
that, a failure to read or parse the failure injection config should
never result in failure to start-up. Hence, all errors are logged and
ignored.