vim: Add search paths for coc-clangd plugin
Instead of symlinking compilation_commands.json from the source
directory into the root, we can configure the clangd plugin to search
in build directories. This removes a manual step and supports release or
debug builds. We default to searching for the file in debug build
directory as we expect most folks default to building redpanda in dev
mode.

See full list of configuration options for coc-clangd:
https://github.com/clangd/coc-clangd/tree/master#configurations

Logic for locating a compilation_commands.json file:
https://github.com/clangd/coc-clangd/blob/bb97d96db42f6ed567157fa87113779fbcda7b9d/src/ctx.ts#L165