archival: compile archival into cluster library
The archival and cluster libraries are intertwined to the point that
separating them to resolve a circular dependency has proved quite
difficult. On the other hand, compiling archival into the cluster
library itself fully resolves the issue.

This commit does not move any files, but merely changes the build system
to reflect the new compilation strategy. This is useful because there
are many inflight changes that would incur a large cost due to a file
move and namespace changes. Future clean up should address this
temporary layout.

Signed-off-by: Noah Watkins <noahwatkins@gmail.com>