cloud_storage: Add fallback mode in segment hydration
When a remote segment is hydrated, it may require either the segment and
the index to be hydrated, if the segment meta version is v2 or earlier,
or it may require only the index to be hydrated, if the meta version is
v3 or later.

There are cases where the index may not be available. In this case,
instead of not allowing reads from the segment, this commit introduces a
fallback mode. This enables format v3 or later to use older logic where
the full segment is downloaded and used to generate the index on the fly.

For newer formats, the first attempt is to download the index. If the
download fails with a permanent error, the mode is switched to fallback
and download is attempted again.

Chunk reading is effectively disabled for this segment. This mode is in
effect until the segment is evicted. It may be possible to switch back to
the chunk mode and revert the fallback once the index has been created,
but this is not implemented yet.

These changes do not affect the behaviour with respect to the
transaction manifest which is downloaded and materialized as before.