txn: new coordinator mapper for tm partitions
We use new hash for choosing coordinator partition by tx_id.
It matches partition to specific section of hash value.
tx is considered to be on specific partition if its hash belongs to
section for this partiion.
This hash will allow us later to define sections for partitions
manually, add and delete partitions