cloud_storage: list objects for s3_imposter
This adds handling of list_objects GET requests to the s3_imposter. It
does a basic scan through all expectations, filtering by a possible
prefix, and grouping based on a delimiter.

I didn't do much to make it robust -- things upload timestamps and tags
are completely fake. That said, this is good enough for some basic list
queries.