cloud_storage: remove dependency on raft
The business logic for sleeping and skipping finalization that is baked
into remote_partition::{finalize,erase} depended on raft configuration
state. This resulted in an inconvenient dependency between the two
subsystems. This PR removes this dependency by moving the business logic
for this behavior that depends on the raft state to the user of the
remote partition.

The goal of this transformation is to preserve exactly the same
semantics. A reviewer will likely notice what appears to be two separate
mechanisms for determining if the current node is a voter or not. It is
not obvious that these two mechanisms (APIs) are identical. If so, a
future clean-up may be useful to remove the ambiguity in the the voter
check.

Signed-off-by: Noah Watkins <noahwatkins@gmail.com>