tests: improve action injector shutdown
Previously it could leave a node in state
where the process had been started but hadn't
finished starting up yet.  That caused the
post-test validation code to trip up trying
to access a node's admin API.

In general, it is valuable for our test primitives
to leave the cluster in a well known state, so rather
than tolerate this in the validation, I'm updating
the action injector to wait for all nodes to be fully
ready before returning.

Fixes https://github.com/redpanda-data/redpanda/issues/10087