tree_wide: replace debug_sanitize flag with config
This change removes the `storage::debug_sanitize_files` flag that was
previously used to enable the file sanitizer in unit tests with the
`std::optional<file_sanitize_config>` type which controls failure
injection and sanitizing behaviour.

The failure injection configuration is read from a JSON file on disk at
start-up and propagated from storage API all the way down to the file
handle creation in the storage layer.

A lot of tests had to be updated to replace the old flag, but the
changes were very mechanical.

Note that this commit stops at the inline `wrap_handle` function in
`segment_utils.cc` and does not propagate the configuration to the file
wrapper. This is intentional, in order to not make this commit even
bigger. A future commit updates the file wrapper and uses the
configuration to inject failures.