cloud_storage: Conditional segment stream creation
When a remote segment is operating in legacy mode, IE either the meta
version is v2 or lower OR fallback mode is engaged, when a reader asks
for a stream it is given the normal file input stream connected to the
segment file.

When the remote segment is operating in chunk mode, a reader is given a
custom data source which hydrates chunks as required while reading the
offsets specified as part of the fetch request.