k/group: log error replicating special records with warn severity
Error replicating cleanup/delete tombstone record is not a critical
error that we should be logging with `ERROR` severity. In future we may
add a retry mechanism in case when replication failed. For now we treat
it as a best effort.

Fixes: #9839

Signed-off-by: Michal Maslanka <michal@redpanda.com>
(cherry picked from commit 8da20402e3eac035217fa8a4185a19fc23184b92)