cloud_storage: Add chunk data source
A chunk data source is added which controls readers accessing chunk files.
The data source manages hydration of chunks when required, and switching
streams from one chunk to the next.

Additionally the segment_chunks API is adjusted to track metrics per
chunk to be able to decide which chunks to trim out of the hydrated list.