r/offset_translator: fix state assignment in offset translator
When `raft::offset_translator` updated state during
`prefix_truncate_reset` all the users which cached the `_state` pointer
wouldn't see the state updates as the assignment didn't change the
underlying data structure but the pointer itself. Fixed the issue by
using an assignment operator that changes pointee not the pointer.

Signed-off-by: Michal Maslanka <michal@redpanda.com>