[storage] Close stream on leadership changes
Here is the situation:

We're uploading a segment and as we start we lose leadership.
`ntp_archiver::upload_segment` is called. Then we call
`do_upload_segment` with `stream_reference`. Within
`ntp_archiver::do_upload_segment` we call `_remote._upload_segment`. We
acquire the lease, then the `lazy_abort_source` requests an abort. We
return normally and the stream ref is still held because we never made
it around to taking the stream from `stream_reference` and nobody closes
the stream (we only close the stream in exceptional cases). This causes
the assertion to be triggered when the stream is destroyed.

The proper thing is to always call `maybe_close_ref()` because if an
upload actually was attempted, then the stream reference will be taken.

Fixes: #10242