config: add a list of tput-controlled API keys
For the purpose of cluster/node-wide tput limits, add a list with API
key nos that will be accounted for and that will be throttled. That would
allow for fine tuning the scope of the feature. The default setting is
Produce (0) and Fetch (1). This may be a temporary property, if we gain
enough confidence that there is a setting (hopefully the default) that
fits all practical cases, the property can be removed and the list can
become hardcoded.