k/quotas: introduce api_key scope for the SNC quotas
SNC quotas are now only accounted for the api_keys listed in the
`kafka_throughput_controlled_api_keys` property. Also only the requests
with keys from that list are subject to throttling because of SNC
quota violation.

`kafka_throughput_controlled_api_keys` is contains API key names.
It is bound with `conversion_binding` to a `vector<bool>` indexed by
API key value. The vector is always populated for the entire range of
API key values so that there is no need to check its size when looking
up for permitted API keys in runtime.