storage: refine retention_adjust_timestamps
Previously, this was adjusting segments with
bogus max_timestamp on each call.  That works,
but is inefficient, as it will mean redundantly
re-adjusting timestamps as long as a segment
with a bogus timestamp is ahead of the retention
time cutoff (on each housekeeping call).

Instead, compare to retention_timestamp(), so that
if we already adjusted the timestamp on a segment,
we will see a valid TS and drop out of our loop.