rptest: fix cloud start kafka offset computation
Previously, `BucketView.kafka_start_offset` computed the start Kafka
offset from the manifest from the segment with the min `base_offset`.
Since the manifest is allowed to have entries in 'segments' that are
below 'start_offset', it broke when, in a recent change, we stopped
uploading after the housekeeping.

This commit fixes the failures in `CloudStorageTimingStressTest` by
correcting the util function mentioned above. Cloud retention always
operats on segment boundaries. There are two cases:
1. Start offset is advanced to the `base_offset` of another segment
2. Start offset is advanced to the `committed_offset + 1` of the latest
   segment at the time. This happens when retention decides to remove
   all segments from the cloud to satisfy the retention policy.

Both cases are now handled correctly.