http/tests: Add ability to store headers per request
Sometimes tests need to inspect the headers which were added to a
request by redpanda. The request info object is made to carry a hashmap
of the headers which a test will be interested in. The test can
register interest by adding the headers to the call to set expectations
and listen.

This feature is only added to s3 imposter for now.