c/bootstrap_backend: set feature table applied offset
Previously, when we fast-forwarded the feature table on applying the
bootstrap command, we didn't set the applied offset of the feature
table. This could trip up the check in the code that applies a
controller snapshot and compares the applied offset of the feature table
and the snapshot to decide if we need to apply the snapshot.