ssx: add a logging_rwlock
This commit adds a wrapper around ss::rwlock that logs when lock
acquisition takes longer than a default of 10s. It is meant to be a
drop-in replacement, though this commit only adds just enough of the
ss::rwlock API to replace current usages in the context of segment
locks.

The following logging will occur:
- trace message is logged when lock attempt is started
- trace message is logged when lock is acquired
- trace message is logged when lock is dropped
- error message is logged when a lock takes longer than 10s to acquire

The ss::rwlock::holder abstraction is also copied with a
ssx::logging_rwlock::holder, which includes a holder ID that is also
logged, allowing for easier correlation between log messages.

This will be useful in debugging deadlocks among the storage layer.