cloud_storage: Remove bucket_not_found error
The error is handled the same way as normal 'fail' outcome and it's
printed as 'unknown' error in the logs (the case in operator << is
missing). The 'remote' api is supposed to give only three error
outcomes. 'retry' in case of retriable error, 'fail' in case of
non-retriable error, and 'no_such_key' if the object is missing from S3.
Every outcome is actionable in unique way but the 'bucket_not_found' is
always handled the same way as 'fail'.