cloud_storage: Remove 'none' error outcome
It's never used and the only handling strategy for it is an assertion.
The actual outcome value is never set to none by s3/abs clients. And it
doesn't make much sense to have it in the first place because there is
no valid way to use it with the 'result<T, cloud_storage::error_outcome>'.