c/topic_table: fix metrics calculation when applying controller snapshot
Previously, the in_progress_update object was constructed with with the
reconfiguration state taken from the snapshot. The problem with this is
that we miss the metrics update associated with the in_progress ->
cancelled transition. Fix this by only allowing constructing in_progress_update
with the the initial in_progress state.