cluster: update topic config used by archiver
`cluster::partition` caches a copy of the current topic configuration.
This cached copy is read by the archiver in order to reupload the topic
manifest when notified by `cluster::partition`.

Prior to this commit, the cached copy was not updated when the
configuration of the topic changed. This meant that the entries in the
topic manifest would not be updated to reflect the change within the
cluster. This could be problematic during recovery:
1. User creates a topic with a small `retention.bytes` value
2. User increses `retention.bytes`. Topic manifest does not get updated
3. Something goes wrong
4. User recovers topic via `redpanda.remote.recovery`. The re-created
   topic will set it's `retention.bytes` to the value provided in step
   1, which is obviously not good.

(cherry picked from commit 1114aa9d760ae0e2ba8f97f69402bef6c960ec64)