[ntp_archiver] Simplify upload_context
Currently upload_context has a method `schedule_single_upload` and field
`uploads` that complicate the data flow and require moving
scheduled_uploads in and out of the uploads array.
`upload_context::scheduled_single_upload` can just be inlined at the
callsite for clearer semantics of what's going on (bumping the offset,
handling of the single returned upload). This also makes error handling
easier as now we just need to cleanup the scheduled_uploads vector if
there is an exception.