tests: make cloud storage movement deletion test more flexible
This previously required exact segment equality, but that is
incompatible with having a low interval between housekeeping,
because adjacent segment merging can kick in and result in
some previously-existing segments no longer existing.

Use the partition size bytes instead of the segment list as
the success condition, and rely on the generic end-of-test
checks to ensure that the size reported in the manifest corresponds
to segments that really exist.