tests: clarify CloudRetentionTimelyGCTest
This test had a rather odd use of wait_until where
the test expected its success condition never to
be met.

The cloud_log_size function was also returning a size
that was never used, so clarify that by calling it
check_cloud_log_size, as its only function is to
raise or not raise an exception.