cloud_storage: Add coarse grained map helper to index
The index can produce a coarse grained (lower resolution than the
original index) map of kafka offsets to file positions. This map can
then be used to find which byte range of a segment a given kafka offset
may lie in.

Because entries are only added to the index at batch start, using this
map makes sure that batch boundaries are not compromised, because the
file offsets here correspond to the start of a batch.