c/utils: validate resulting configuration only in terms of changed field
When cluster member configuration changes Redpanda by default doesn't
allow the change if a new cluster configuration would have two
listeners pointing to the same address. This was in conflict with the
logic of join request which didn't execute validation of resulting
configuration. Change the validation logic to only check configuration
fields which were changed and are going to be updated.

Fixes: #9602

Signed-off-by: Michal Maslanka <michal@redpanda.com>