Reverse the comparator order in topic_partition
For the == operator for topic partition, compare the partition first,
then the topic. This is beneficial since when the objects are different
it is more likely that the partition varies than the topic, since
there are in general more partitions than topics in a system (actually
we require that there are more unique partition IDs than topics, which
is a bit less likely but still seems very likely on average).

Furthermore, this comparison is very cheap, so even if the likelihoods
were similar it makes sense to do it first. This is relevant, for
example, when looking up these objects in a hash map, since non-equal
comparisons occur when there are multiple objects in one bucket.

Issue redpanda-data/core-internal#435