Introduce ktp class
A ktp is an ntp in the kafka namespace.

The namespace is implicitly kafka, but is not stored, for
performance reasons. I.e., it logically represents the 3-tuple
("kafka", <topic>, <partition>) but only the (<topic>, <partition>)
part is stored.

Avoiding storing the namespace makes, creating, copying, hashing and
equality checking (among others) considerably faster. For example,
hashing a ktp object with a 30-characer topic name takes ~67 cycles,
versus 117 cycles for the equivalent ntp object. The relative difference
is larger for shorter topic names.

This class can be compared with an ntp and will be equal if the
3-tuple is equal. Similarly it can be hashed with the same result
as an ntp in the kafka namespace. This means you can replace a use
of ntp with ktp for topics in the kafka namespace, but still inter-
operate with ntp objects, e.g., as keys in a hash table (a transparent
comparator is needed).

This change introduces the ktp class in its own header as well as
a ktp unit test which checks equivalence of the relevant operators with
the ntp class.

In addition, a ktp_with_hash class is added, which is like ktp but
memorizes its own hash as a field, so that hashing is very fast.

Issue redpanda-data/core-internal#435