r/hb_manager: do not drop connections on logical hb errors
In Raft we have a mechanism that drops the connection if there is a
particular number of failed heartbeats. The connection should only be
dropped if it is unresponsive. Currently all the errors in heartbeat
responses result in a connection being dropped.

Changed the implementation to only mark heartbeats as failed if the
whole request failed with timeout. This way a connection will only be
dropped when the node is not responsive.

Signed-off-by: Michal Maslanka <michal@redpanda.com>