Extend hash_bench to include ntp hashes
Fixed up the old hash_bench to do a number of iterations in the inner
loop since it was returning useless results currently due to the
high overhead of PERF_TEST start/stop (it uses perf to get instruction)
counts which takes 100s of cycles.

Added ntp and ktp hash methods to validate the improvement provided
by the ktp class.

Issue redpanda-data/core-internal#435