Don't use auto_measure for fetch latency probe
We created one of these for every partition associated with a fetch
request, which may be 100s or 1000s of partitions, many of which will
never return any bytes for the current request. Since auto_measure
involves an allocation, this results in many spurious allocations on
the fetch path.

Since we only really need the start timestamp, just track and and
record it in the histogram manually.

Issue redpanda-data/core-internal#435