config: support nonintegral bound properties
`numeric_bounds<T>` implied the argument to be an integral numeric
by requiring the `%` operation on it. This change renames `numeric_bounds`
into `numeric_integral_bounds` to emphasize that, and introduces
the `numeric_bounds` that does not support alignments and odd/even checks,
and thus works with floating point types too.

The `bounded_property` now can accept an arbitrary bounds struct that
conforms to `detail::bounds<>` concept. For compatibility purpose,
it defaults to `numeric_integral_bounds` so no code change is necessary.