[raft] Streaming serde for record batches
This collection during the
`ScalingUpTest/adding_multiple_nodes_to_the_cluster/partition_count=20`
there is a point at which we attempt to write out a batch of 1193
elements, which triggers our oversized allocation warning. Stream out
the records instead and prepend the size header into iobuf after the
fact.

I still need to fix that we allocate a large batch in stm_recovery as
the input to this serde, but that will come in another change.

Fixes: https://github.com/redpanda-data/redpanda/issues/10303