archival: avoid lock inversion in segment merger
The main archival loop currently takes locks in the order 1)
ntp_archiver_service::_mutex, 2) segment locks. The segment merging
housekeeping job takes locks in the opposite order.

This could result in a deadlock when attempting to take a lock on the
same segment if the segment merger attempts to merge a partially
uploaded segment, and concurrently the archival loop attempts to do
another partial segment upload of that segment.

This commit extends the ntp_archiver_service interface to force callers
of ntp_archival_service::upload() to already have a lock, rather than
acquiring the lock therein.

Fixes #10085