application: wait for cluster ID before starting Kafka API
Previously, we would expose a `redpanda.initializing` cluster ID
in the period between the kafka API starting, and the metrics reporter
initializating the cluster UUID.

Now that the metrics reporter starts fast, we may block on
the cluster ID being set to start listening on the
kafka API.

This assumes that enough batches are written to the controller
log to initialize the cluster UUID before we listen for
kafka requests for the first time.  This is true in
practice (bootstrap, config, features batches).