tests/redpanda: Don't exclude redpanda with ']' in args
A change was introduced in 81c37c5 that attempted to fix an issue where
an incorrect process was selected as the redpanda process.  The
unintended side-effect was that any actual Redpanda process that
contained a ']' was excluded so methods like `self.redpanda.stop()`
wouldn't actually stop Redpanda as it was unable to find the appropriate
PID.

Fixed `ps` line to exclude "[redpanda]" rather than just "]".

Signed-off-by: Michael Boquard <michael@redpanda.com>