cluster: ensure prompt propagation of cluster ID from metrics reporter
Previously, if we initialized the cluster ID before a raft0 leader
was elected, we would drop out of do_report_metrics before setting
cluster_id in the global configuration.

This would cause a delay of up to the metrics reporter interval
before the Kafka API can start.

Instead, spin waiting for raft leadership or for the cluster_id
to be set.