c/partition_allocator: use chunked_fifo to store allocation units
When dealing with a topic that has a lot of partitions all the
partition allocations are gathered in a single collection. Previously
the allocations were stored in a `std::vector` this lead to a situation
in which a vector was large enough to trigger oversized allocation
warning. Replaced a vector with `ss::chunked_fifo` to prevent large
allocations.

Fixes: #10350
Fixes: #10351

Signed-off-by: Michal Maslanka <michal@redpanda.com>
(cherry picked from commit eefb1306f60e6b5cf34c914e99197a206b35a270)