tests: update test_misconfigured_root_driven_bootstrap
This test previously asserted that nodes which could not
form a cluster but which had node_ids set would be listening
on their admin APIs.  That is no longer the case, now that
such nodes insist on doing a join_node_request to the cluster
to get a snapshot before they proceed with starting runtime
services.