storage: add more cached size clearing
There were a few cases in which cached segment size was not accurate,
such as when appender would fallocate additional capacity or compaction
would swap out the underlying file of a segment. This lead to cases in
which the size reported was significantly smaller than the on disk size
observed during testing.

Signed-off-by: Noah Watkins <noahwatkins@gmail.com>