test: add storage usage verification on test shutdown
This patch adds a storage usage verification step to the shutdown
process to every test (similar to bad log lines). the threshold is a 5%
difference. Currently we ignore cloud_storage_cache in the calculation,
as well as a few files like pidflie and config cache.

The 5% threshold can be lowered in the future to be even tighter.

Signed-off-by: Noah Watkins <noahwatkins@gmail.com>