admin_server: Add sampled_memory_profile endpoint
Adds an endpoint that the collects the current sampled memory profile
from all shards and returns them to the caller.

For now the stack is just stringified but we could make this a proper
json structure all the way down. Keeping it simple for now.

The following (or some form thereof) can be used to get a flamegraph:

```
curl localhost:9644/v1/debug/sampled_memory_profile?shard=3 \
    | jq -r .[0].profile \
    | ./symbolize_mem_dump.py /home/stephan/build/redpanda/vbuild/release/clang/bin/redpanda \
    | flamegraph.pl - > flamegraph.svg
```