storage: drop err log on flush with concurrent ops
The error injection infra inherited an error log from
`file_io_sanitizer` which triggered when a flush would be triggered
concurrently with any other operation on the segment appender.

It turns out that this is safe in the general case. The segment appender
can perform two types of flushes:
1. soft flush via segment_appender::flush() - These flushes can be
   concurrent with writes past the "flushing ofset"
2. hard flushes via segment_appender::hard_flush() - Theses flushes
   should never be concurrent with writes. This is in enforced by
   assertions within the segment appender.

Case 1 needs a bit more context. Each call to flush within the appender
is associated with an offset to flush at, called "flushing offset"
previously. It's computed as the last offset we've written at plus the
pending bytes. The appender enforces the invariant that all write calls
to offsets before the flushing offset have completed. Hence, flushes
being concurrent with writes past the "flushing offset" are safe, as the
application does not believe it flushed when it has not actually done
so.