storage: drop clean_segment record on truncate
Previously, we could be partway through truncating
a previously cleanly-shutdown segment, and have written
out indices but not data files, and thereby violate
the invariant that a "clean" segment has all its data
and indices on disk in a consistent state.

This would be unlikely to cause immediate issues, but
could be a serious problem if we e.g. later delete the
truncated indices and it gets rebuilt from the un-truncated
data file.