schema_registry/test: Support NestedPayload
The serde format for Protobuf has an encoding for the Message inside
the protobuf file. The encoding is a length, followed by offsets into the
file and messages until the correct nested message is found. If the Message
is the first message in the file, the encoding is just a 0 for the length.

Protobuf Payload is used to check the optimised encoding.

Protobuf NestedPayload is used to check the full encoding with length and
offsets. In the case of `com.redpanda.A.B.C.D.NestedPayload, the encoding
is the following sequence of zigzag encoded integers:
[5, 1, 0, 0, 0, 18]

Signed-off-by: Ben Pope <ben@redpanda.com>